working on parsing

+++ common/src/main/java/de/hpi/mmds/wiki/Recommender.java
@@ -6,9 +6,9 @@ import org.apache.spark.api.java.JavaRDD;
 
 public interface Recommender {
 	
-	default List<Recommendation> recommend(int userId, JavaRDD<Long> articles) {
+	default List<Recommendation> recommend(int userId, JavaRDD<Integer> articles) {
 		return recommend(userId, articles, 10);
 	}
 
-	List<Recommendation> recommend(int userId, JavaRDD<Long> articles, int howMany);
+	List<Recommendation> recommend(int userId, JavaRDD<Integer> articles, int howMany);
 }
